+++ Controllers/AzureProviderController.cs
@@ -1,18 +1,16 @@
-﻿using Microsoft.AspNetCore.Cors;
+﻿using System;
+using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
-using Microsoft.AspNetCore.Http.Features;
-using System;
 using Syncfusion.EJ2.FileManager.Base;
 using Syncfusion.EJ2.FileManager.AzureFileProvider;
 
-
 namespace EJ2AzureASPCoreFileProvider.Controllers
 {
-
     [Route("api/[controller]")]
     [EnableCors("AllowAllOrigins")]
     public class AzureProviderController : Controller
@@ -24,7 +22,7 @@ namespace EJ2AzureASPCoreFileProvider.Controllers
             this.operation.RegisterAzure("<--accountName-->", "<--accountKey-->", "<--blobName-->");
             this.operation.SetBlobContainer("<--blobPath-->", "<--filePath-->");
             //----------
-            //for example 
+            //For example 
             //this.operation.RegisterAzure("azure_service_account", "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx", "files");
             //this.operation.setBlobContainer("https://azure_service_account.blob.core.windows.net/files/", "https://azure_service_account.blob.core.windows.net/files/Files");
             //---------
@@ -37,7 +35,7 @@ namespace EJ2AzureASPCoreFileProvider.Controllers
                 string startPath = "<--blobPath-->";
                 string originalPath = ("<--filePath-->").Replace(startPath, "");
                 //-----------------
-                // for example
+                //For example
                 //string startPath = "https://azure_service_account.blob.core.windows.net/files/";
                 //string originalPath = ("https://azure_service_account.blob.core.windows.net/files/Files").Replace(startPath, "");
                 //-------------------
@@ -47,30 +45,29 @@ namespace EJ2AzureASPCoreFileProvider.Controllers
             switch (args.Action)
             {
                 case "read":
-                    // reads the file(s) or folder(s) from the given path.
-                    return Json(this.ToCamelCase(this.operation.GetFiles(args.Path, args.Data)));
+                    // Reads the file(s) or folder(s) from the given path.
+                    return Json(this.ToCamelCase(this.operation.GetFiles(args.Path, args.ShowHiddenItems, args.Data)));
                 case "delete":
-                    // deletes the selected file(s) or folder(s) from the given path.
+                    // Deletes the selected file(s) or folder(s) from the given path.
                     return this.ToCamelCase(this.operation.Delete(args.Path, args.Names, args.Data));
                 case "details":
-                    // gets the details of the selected file(s) or folder(s).
+                    // Gets the details of the selected file(s) or folder(s).
                     return this.ToCamelCase(this.operation.Details(args.Path, args.Names, args.Data));
                 case "create":
-                    // creates a new folder in a given path.
-                    return this.ToCamelCase(this.operation.Create(args.Path, args.Name));
+                    // Creates a new folder in a given path.
+                    return this.ToCamelCase(this.operation.Create(args.Path, args.Name, args.Data));
                 case "search":
-                    // gets the list of file(s) or folder(s) from a given path based on the searched key string.
+                    // Gets the list of file(s) or folder(s) from a given path based on the searched key string.
                     return this.ToCamelCase(this.operation.Search(args.Path, args.SearchString, args.ShowHiddenItems, args.CaseSensitive, args.Data));
                 case "rename":
-                    // renames a file or folder.
+                    // Renames a file or folder.
                     return this.ToCamelCase(this.operation.Rename(args.Path, args.Name, args.NewName, false, args.Data));
                 case "copy":
-                    // copies the selected file(s) or folder(s) from a path and then pastes them into a given target path.
+                    // Copies the selected file(s) or folder(s) from a path and then pastes them into a given target path.
                     return this.ToCamelCase(this.operation.Copy(args.Path, args.TargetPath, args.Names, args.RenameFiles, args.TargetData, args.Data));
                 case "move":
-                    // cuts the selected file(s) or folder(s) from a path and then pastes them into a given target path.
+                    // Cuts the selected file(s) or folder(s) from a path and then pastes them into a given target path.
                     return this.ToCamelCase(this.operation.Move(args.Path, args.TargetPath, args.Names, args.RenameFiles, args.TargetData, args.Data));
-
             }
             return null;
         }
@@ -85,7 +82,7 @@ namespace EJ2AzureASPCoreFileProvider.Controllers
             });
         }
 
-        // uploads the file(s) into a specified path
+        // Uploads the file(s) into a specified path
         [Route("AzureUpload")]
         public ActionResult AzureUpload(FileManagerDirectoryContent args)
         {
@@ -95,14 +92,13 @@ namespace EJ2AzureASPCoreFileProvider.Controllers
                 string originalPath = ("<--filePath-->").Replace(startPath, "");
                 args.Path = (originalPath + args.Path).Replace("//", "/");
                 //----------------------
-                //for example
+                //For example
                 //string startPath = "https://azure_service_account.blob.core.windows.net/files/";
                 //string originalPath = ("https://azure_service_account.blob.core.windows.net/files/Files").Replace(startPath, "");
                 //args.Path = (originalPath + args.Path).Replace("//", "/");
                 //----------------------
             }
-            FileManagerResponse uploadResponse;
-            uploadResponse = operation.Upload(args.Path, args.UploadFiles, args.Action, args.Data);
+            FileManagerResponse uploadResponse = operation.Upload(args.Path, args.UploadFiles, args.Action, args.Data);
             if (uploadResponse.Error != null)
             {
                 Response.Clear();
@@ -113,7 +109,7 @@ namespace EJ2AzureASPCoreFileProvider.Controllers
             return Json("");
         }
 
-        // downloads the selected file(s) and folder(s)
+        // Downloads the selected file(s) and folder(s)
         [Route("AzureDownload")]
         public object AzureDownload(string downloadInput)
         {
@@ -121,7 +117,7 @@ namespace EJ2AzureASPCoreFileProvider.Controllers
             return operation.Download(args.Path, args.Names, args.Data);
         }
 
-        // gets the image(s) from the given path
+        // Gets the image(s) from the given path
         [Route("AzureGetImage")]
         public IActionResult AzureGetImage(FileManagerDirectoryContent args)
         {